Changed resource requests addresses to match conventions.
